\name{KUL3}
\docType{data}
\alias{KUL3}
\title{Colorectal Cancer Aggregated Single Cells from Leuven}
\description{
  This data is an aggregate single-cell summaries derived from
  publicly available raw counts and annotations in GSE144735}
\usage{KUL3}
\format{A list of aggregated annotation data.frame and
  summarized expression array}
\source{Gene Expression Omnibus GSE144735}
\references{
  Lee H-O, Hong Y, Etlioglu HE, Cho YB, Pomella V, Van den Bosch B,
  Vanhecke J, Verbandt S, Hong H, Min J-W, Kim N, Eum HH, Qian J,
  Boeckx B, Lambrechts D, Tsantoulis P, De Hertogh G, Chung W,
  Lee T, An M, Shin H-T, Joung J-G, Jung M-H, Ko G, Wirapati P,
  Kim SH, Kim HC, Yun SH, Tan IBH, Ranjan B, Lee WY, Kim T-Y, 
  Choi JK, Kim Y-J, Prabhakar S, Tejpar S, Park W-Y (2020)
  Lineage-dependent gene expression programs influence the immune
  landscape of colorectal cancer. Nat Genet 52(6): 594-603. [PMID:32451460]
}
\keyword{datasets}
